# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "post API Documentation",
        "version": "0.0.1"
    },
    "paths": {
        "/api/post/insert": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "create post",
                "description": "post create here",
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "create Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "create Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/post/list": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get list of posts",
                "description": "Returns a list of posts.",
                "operationId": "listPosts",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/post/{post}/get": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "get post",
                "operationId": "get",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid post supplied"
                    },
                    "404": {
                        "description": "post not found"
                    }
                }
            }
        },
        "/api/post/{post}/update": {
            "put": {
                "tags": [
                    "post"
                ],
                "summary": "Update post",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the post to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update a post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of the post (optional, min 3 chars)",
                                        "type": "string",
                                        "minLength": 3
                                    },
                                    "description": {
                                        "description": "Description of the post (optional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid post supplied"
                    },
                    "404": {
                        "description": "post not found"
                    }
                }
            }
        },
        "/api/post/{post}/delete": {
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "delete post",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "ID of the post to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid post supplied"
                    },
                    "404": {
                        "description": "post not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Post": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "post",
            "description": "post"
        }
    ]
}